+++ OrcCreator/OrcCreator.vcxproj
@@ -139,6 +139,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="orc.cpp" />
     <ClCompile Include="OrcCreator.cpp" />
   </ItemGroup>
   <ItemGroup>
