+++ OrcCreator/OrcCreator.vcxproj.filters
@@ -18,6 +18,9 @@
     <ClCompile Include="OrcCreator.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="orc.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="orc.h">
